requirements.txt
beautifulsoup4==4.14.2
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-Mail==0.10.0
Flask-Compress==1.15
Pillow==10.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
soupsieve==2.8
typing_extensions==4.15.0
Werkzeug==3.1.3